# docs/swagger.yaml
definitions:
...
  apiSponsor.ChannelAddRequest:
    properties:
      description:
        description: 渠道描述
        example: 支付宝赞助渠道
        maxLength: 1000
        type: string
      icon:
        description: 渠道图标 URL
        example: https://example.com/icon.png
        maxLength: 500
        type: string
      name:
        description: 渠道名称
        example: 支付宝
        maxLength: 50
        minLength: 1
        type: string
      sort_order:
        description: 排序值
        example: 0
        minimum: 0
        type: integer
    required:
    - name
    type: object
  apiSponsor.ChannelAddResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      sort_order:
        type: integer
      sponsor_count:
        type: integer
      status:
        type: boolean
      updated_at:
        type: string
    type: object
  apiSponsor.ChannelDeleteResponse:
    properties:
      message:
        type: string
    type: object
  apiSponsor.ChannelDetailResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      sort_order:
        type: integer
      sponsor_count:
        type: integer
      status:
        type: boolean
      updated_at:
        type: string
    type: object
  apiSponsor.ChannelEntityResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      sort_order:
        type: integer
      sponsor_count:
        type: integer
      status:
        type: boolean
      updated_at:
        type: string
    type: object
  apiSponsor.ChannelListItemResponse:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      sort_order:
        type: integer
      sponsor_count:
        type: integer
      status:
        type: boolean
    type: object
  apiSponsor.ChannelPageResponse:
    properties:
      data:
        description: 数据列表
        items:
          $ref: '#/definitions/apiSponsor.ChannelEntityResponse'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/base.PaginationInfo'
        description: 分页信息
    type: object
  apiSponsor.ChannelStatusRequest:
    properties:
      status:
        description: 状态:true=启用,false=禁用
        example: true
        type: boolean
    type: object
  apiSponsor.ChannelStatusResponse:
    properties:
      message:
        type: string
      status:
        description: 更新后的状态
        type: boolean
    type: object
  apiSponsor.ChannelUpdateRequest:
    properties:
      description:
        description: 渠道描述
        example: ""
        maxLength: 1000
        type: string
      icon:
        description: 渠道图标 URL
        example: https://example.com/icon.png
        maxLength: 500
        type: string
      name:
        description: 渠道名称
        example: 支付宝
        maxLength: 50
        minLength: 1
        type: string
      sort_order:
        description: 排序值
        example: 0
        minimum: 0
        type: integer
    type: object
  apiSponsor.ChannelUpdateResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      sort_order:
        type: integer
      sponsor_count:
        type: integer
      status:
        type: boolean
      updated_at:
        type: string
    type: object
...
  base.PaginationInfo:
    properties:
      has_next:
        description: 是否有下一页
        type: boolean
      has_prev:
        description: 是否有上一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/sponsors/channels:
    get:
      consumes:
      - application/json
      description: 分页获取赞助渠道列表，支持过滤和排序
      parameters:
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10，最大100）
        in: query
        name: page_size
        type: integer
      - description: 状态过滤（true=启用，false=禁用）
        in: query
        name: status
        type: boolean
      - description: 名称模糊搜索
        in: query
        name: name
        type: string
      - description: 排序字段（name, sort_order, created_at）
        in: query
        name: order_by
        type: string
      - description: 排序方向（asc, desc）
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/apiSponsor.ChannelPageResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取赞助渠道分页列表
      tags:
      - 赞助渠道管理
    post:
      consumes:
      - application/json
      description: 创建新的赞助渠道
      parameters:
      - description: 添加赞助渠道请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiSponsor.ChannelAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            $ref: '#/definitions/apiSponsor.ChannelAddResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 添加赞助渠道
      tags:
      - 赞助渠道管理
  /api/v1/admin/sponsors/channels/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的赞助渠道
      parameters:
      - description: 赞助渠道ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/apiSponsor.ChannelDeleteResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 赞助渠道不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 删除赞助渠道
      tags:
      - 赞助渠道管理
    get:
      consumes:
      - application/json
      description: 根据ID获取指定赞助渠道的详细信息
      parameters:
      - description: 赞助渠道ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/apiSponsor.ChannelDetailResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 赞助渠道不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取赞助渠道详情
      tags:
      - 赞助渠道管理
    put:
      consumes:
      - application/json
      description: 更新指定赞助渠道的信息
      parameters:
      - description: 赞助渠道ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新赞助渠道请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiSponsor.ChannelUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/apiSponsor.ChannelUpdateResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 赞助渠道不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 更新赞助渠道
      tags:
      - 赞助渠道管理
  /api/v1/admin/sponsors/channels/{id}/status:
    patch:
      consumes:
      - application/json
      description: 切换指定赞助渠道的启用/禁用状态
      parameters:
      - description: 赞助渠道ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 渠道状态请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiSponsor.ChannelStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 状态更新成功
          schema:
            $ref: '#/definitions/apiSponsor.ChannelStatusResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 赞助渠道不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 更新赞助渠道状态
      tags:
      - 赞助渠道管理
  /api/v1/admin/sponsors/channels/all:
    get:
      consumes:
      - application/json
      description: 获取赞助渠道列表（不分页），支持过滤和排序
      parameters:
      - description: 状态过滤（true=启用，false=禁用）
        in: query
        name: status
        type: boolean
      - description: 名称模糊搜索
        in: query
        name: name
        type: string
      - description: 仅查询启用的渠道
        in: query
        name: only_enabled
        type: boolean
      - description: 排序字段（name, sort_order, created_at）
        in: query
        name: order_by
        type: string
      - description: 排序方向（asc, desc）
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              $ref: '#/definitions/apiSponsor.ChannelListItemResponse'
            type: array
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取赞助渠道列表
      tags:
      - 赞助渠道管理
...
  /api/v1/sponsors/channels:
    get:
      consumes:
      - application/json
      description: 获取所有启用状态的赞助渠道（不分页，公开接口）
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              $ref: '#/definitions/apiSponsor.ChannelListItemResponse'
            type: array
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取公开的赞助渠道列表
      tags:
      - 赞助渠道
...
swagger: "2.0"